Kha is not an overbearing framework. Calls to `update()` are scheduled to occur 60 times per second, and the `render()` call is logged as a callback for the `System` to execute when appropriate. This is about as little of a footprint Kha can leave in the codebase, which makes it easier to focus on the actual code that needs to be written.
Pong is a simple game. Two paddles on the screen bounce a single ball between them until one of the paddles misses the ball. `Paddle` and `Ball` are distinct objects here. They have some overlap at the fundamental level (position properties, update and draw methods), but for the sake of simplicity are kept in their own classes.

It should be noted that Kha's `Framebuffer` is drawn to via a graphics API, in this case `graphics2.Graphics`. This is a straightforward API for 2D graphics and provides the necessary primitive drawing functions (namely `fillRect()`). A reference to the framebuffer - as accessed by the `graphics2` API - is passed to the `draw()` method of both paddles and the ball, after the framebuffer is set up to draw white primitives onto a black background.

`Ball` and `Paddle` classes are simple. They provide `update()` and `draw()` methods for the main `Project` class to use. However, they draw their visuals in a slightly unconventional manner. The graphic is centered in the object's (x, y) position, meaning that the e.g. ball wouldn't be drawn at (0, 0) starting from the top-left, but rather from (0 - radius, 0 - radius) so that the ball graphic's centre is located at (0, 0).