(Added interaction: paddles are movable by "w" and "s" keys, ball moves around the playing field and collides with paddles.)
Kha handles keyboard input through callback functions which require `kha.Key` and `String` arguments. The `kha.Key` enum provides access to special keyboard keys (escape, backspace, enter, ...), while the `String` argument is the actual character that triggered the event callback. Switching against that character makes it possible to have keyboard interaction with the program.

Basic collision handling requires hit tests against a rectangular collision area. For the ball, colliding against the window bounds is simple. A check is performed for the left and right sides of the window in order to change the horizontal direction the ball is traveling, and another check is performed for the top and right sides of the window to change the vertical direction. The starting direction the ball will travel is determined by a random number generator.

Paddles perform hit tests against the ball in a similar fashion. Left, right, top, and bottom coordinates of the collision bounds are calculated from paddle and ball positions and sizes. When the coordinate ranges overlap, the hit test returns true. However, the ball travels at a certain velocity, which is in a certain sense a "sample rate" for the hit test. A greater velocity results in a smaller sample rate and imprecise collision detection. A solution would be to "check ahead" for the collision and - if it is to occur - move the ball up to the position at the time of collision before introducing collision effects.