(Added simple two-player system. Player 1 plays with "w" and "s", player 2 with "o" and "l".)
Handling input for two players can be implemented in a manner of different ways. The simplest and least elegant way is to branch the switch statements for the key handlers according to the player the paddle is assigned to. A `Player` enum is used to differentiate between players controlling the paddles. Player 1 uses "w" and "s" to control his paddle, while player 2 uses "o" and "l".