(Added rudimentary collision detection.)
Adding simple collision detection is easy now that the `Entity` class provides methods to detect overlaps with points or other entities.

If the ball - after its position has been updated - overlaps a paddle, change its horizontal direction of movement. Then check which section of the paddle has been overlapped to change the vertical direction of movement.

This latter part of the collision detection logic can and should be done differently. The ball should not dictate how the paddle is sectioned, instead the paddle should define its sections internally and make it possible for the ball to find out which section was overlapped. In the spirit of simplicity, however, this implementation will (for now) be kept as it is.