(Added font asset. Added score display. Restructured project.)
In order to use assets with Kha they naturally need to be loaded first. Therefore, before creating a new instance of `Project` (the game), initialisation will take care of loading all assets with a call to `kha.Assets.loadEverything()`. This method takes a callback function for its argument, which is executed once all assets have been readied by Kha.

A font asset is required to write (draw) strings of text such as player scores. The font is loaded into a `font` variable and the font size is set during the rendering process. Scores are kept for both the left and right paddle (players one and two) as an integer. Their string form is then drawn to the framebuffer for the players to see.

As a safety measure, all game-related classes were moved under the `game` namespace.